Rename utils module to map-utils and update import path in main.js
